Finished jwt authentication

+++ server/controllers/employeeController.js
@@ -1,10 +1,13 @@
 const Employee = require("../models/Employee");
-const mongoose = require('mongoose'); 
+const mongoose = require("mongoose");
 
 const getEmployees = async (req, res) => {
   try {
+    const { userId } = req;
+    if (!userId)
+      return res.json({ response: "You need to sign in to your workshop" });
     const employees = await Employee.find(
-      { active: true },
+      { active: true, creator: userId },
       {
         name: 1,
         phoneNumber: 1,
@@ -12,10 +15,10 @@ const getEmployees = async (req, res) => {
         startDate: 1,
         salary: 1,
         address: 1,
+        creator: 1,
       }
     );
-    if (!employees.length)
-      return res.json({ response: [] });
+    if (!employees.length) return res.json({ response: [] });
     res.json({ response: employees });
   } catch (e) {
     console.log(e);
@@ -23,11 +26,12 @@ const getEmployees = async (req, res) => {
 };
 
 const showEmployee = async (req, res) => {
-  const {id} = req.params
-  console.log(id)
+  const { id } = req.params;
   try {
     if (!mongoose.Types.ObjectId.isValid(id))
-    return res.status(404).json({response: "No post with that id sorry 🤦‍♂️"});
+      return res
+        .status(404)
+        .json({ response: "No post with that id sorry 🤦‍♂️" });
 
     const employees = await Employee.find(
       { active: true, _id: id },
@@ -40,9 +44,8 @@ const showEmployee = async (req, res) => {
         address: 1,
       }
     );
-    if (!employees.length)
-      return res.redirect('back')
-      
+    if (!employees.length) return res.redirect("back");
+
     res.json({ response: employees });
   } catch (e) {
     console.log(e);
@@ -52,12 +55,21 @@ const showEmployee = async (req, res) => {
 const addEmployee = async (req, res) => {
   try {
     const { postData } = req.body;
+    const { userId } = req;
     if (!postData)
       return res.json({ error: "You are not allowed to create an empty post" });
-    const employee = await new Employee(postData);
+    const employee = await new Employee({ ...postData, creator: userId });
     await employee.save();
-    const { active, name, phoneNumber, addedDate, startDate, salary, _id } =
-      employee;
+    const {
+      active,
+      name,
+      phoneNumber,
+      addedDate,
+      startDate,
+      salary,
+      _id,
+      creator,
+    } = employee;
     res.json({
       response: {
         active,
@@ -67,6 +79,7 @@ const addEmployee = async (req, res) => {
         startDate,
         salary,
         _id,
+        creator,
       },
     });
   } catch (error) {
@@ -77,13 +90,21 @@ const addEmployee = async (req, res) => {
 const editEmployee = async (req, res) => {
   try {
     const { id } = req.params;
+    console.log(id);
 
+    const { userId } = req;
+
+    const foundEmployee = await Employee.findById(id);
+    if (userId != foundEmployee.creator) {
+      return res.json({
+        response: "You aren't allow to edit if you are not logged in",
+      });
+    }
     const employee = req.body.postData;
-    console.log(employee);
     const updatedEmployee = await Employee.findByIdAndUpdate(id, employee, {
       new: true,
     });
-    // console.log(updatedEmployee)
+    console.log(updatedEmployee);
     res.json({ updatedEmployee });
   } catch (error) {
     console.log(error);
@@ -91,14 +112,21 @@ const editEmployee = async (req, res) => {
 };
 
 const deleteEmployee = async (req, res) => {
-  const {id} = req.params; 
-  await Employee.findByIdAndDelete(id);  
-  res.send('Successfully deleted the post')
+  const { id } = req.params;
+  const { userId } = req;
+  const employee = await Employee.findById(id);
+  if (userId !== employee.creator)
+    return res.json({
+      response: "You aren't allow to edit if you are not logged in",
+    });
+
+  await Employee.findByIdAndDelete(id);
+  res.send("Successfully deleted the post");
 };
 module.exports = {
   getEmployees,
   addEmployee,
   editEmployee,
   deleteEmployee,
-  showEmployee
+  showEmployee,
 };
